+++ lib/massive_record/orm/relations/proxy/references_many.rb
@@ -45,19 +45,19 @@ module MassiveRecord
           # Adding record(s) to the collection.
           #
           def <<(*records)
-            save_records = owner.persisted?
+            save_records = proxy_owner.persisted?
 
             if records.flatten.all? &:valid?
               records.flatten.each do |record|
                 unless include? record
                   raise_if_type_mismatch(record)
-                  add_foreign_key_in_owner(record.id)
+                  add_foreign_key_in_proxy_owner(record.id)
                   target << record
                   record.save if save_records
                 end
               end
 
-              owner.save if save_records
+              proxy_owner.save if save_records
 
               self
             end
@@ -137,26 +137,26 @@ module MassiveRecord
           def delete_or_destroy(*records, method)
             records.flatten.each do |record|
               if include? record
-                remove_foreign_key_in_owner(record.id)
+                remove_foreign_key_in_proxy_owner(record.id)
                 target.delete(record)
                 record.destroy if method.to_sym == :destroy
               end
             end
 
-            owner.save if owner.persisted?
+            proxy_owner.save if proxy_owner.persisted?
           end
 
 
 
           def find_target
-            target_class.find(owner.send(foreign_key), :skip_expected_result_check => true)
+            target_class.find(proxy_owner.send(foreign_key), :skip_expected_result_check => true)
           end
 
           def find_first
             if can_find_target?
               if find_with_proc?
                 find_target_with_proc(:limit => 1).first
-              elsif first_id = owner.send(foreign_key).first
+              elsif first_id = proxy_owner.send(foreign_key).first
                 target_class.find(first_id)
               end
             end
@@ -167,30 +167,30 @@ module MassiveRecord
           end
 
           def can_find_target?
-            super || (owner.respond_to?(foreign_key) && owner.send(foreign_key).any?)
+            super || (proxy_owner.respond_to?(foreign_key) && proxy_owner.send(foreign_key).any?)
           end
 
 
           
 
 
-          def add_foreign_key_in_owner(id)
-            if owner.respond_to? foreign_key
-              owner.send(foreign_key) << id
-              notify_of_change_in_owner_foreign_key
+          def add_foreign_key_in_proxy_owner(id)
+            if proxy_owner.respond_to? foreign_key
+              proxy_owner.send(foreign_key) << id
+              notify_of_change_in_proxy_owner_foreign_key
             end
           end
 
-          def remove_foreign_key_in_owner(id)
-            if owner.respond_to? foreign_key
-              owner.send(foreign_key).delete(id)
-              notify_of_change_in_owner_foreign_key
+          def remove_foreign_key_in_proxy_owner(id)
+            if proxy_owner.respond_to? foreign_key
+              proxy_owner.send(foreign_key).delete(id)
+              notify_of_change_in_proxy_owner_foreign_key
             end
           end
 
-          def notify_of_change_in_owner_foreign_key
+          def notify_of_change_in_proxy_owner_foreign_key
             method = foreign_key+"_will_change!"
-            owner.send(method) if owner.respond_to? method
+            proxy_owner.send(method) if proxy_owner.respond_to? method
           end
         end
       end