+++ lib/massive_record/orm/relations/proxy/references_one.rb
@@ -8,28 +8,28 @@ module MassiveRecord
         class ReferencesOne < Proxy
 
           def target=(target)
-            set_foreign_key_in_owner(target.id) if target
+            set_foreign_key_in_proxy_owner(target.id) if target
             super(target)
           end
 
           def replace(target)
             super
-            set_foreign_key_in_owner(nil) if target.nil?
+            set_foreign_key_in_proxy_owner(nil) if target.nil?
           end
 
 
           private
 
           def find_target
-            target_class.find(owner.send(foreign_key))
+            target_class.find(proxy_owner.send(foreign_key))
           end
 
           def can_find_target?
-            super || owner.send(foreign_key).present?
+            super || proxy_owner.send(foreign_key).present?
           end
 
-          def set_foreign_key_in_owner(id)
-            owner.send(foreign_key_setter, id) if owner.respond_to?(foreign_key_setter)
+          def set_foreign_key_in_proxy_owner(id)
+            proxy_owner.send(foreign_key_setter, id) if proxy_owner.respond_to?(foreign_key_setter)
           end
         end
       end