+++ spec/orm/relations/proxy/references_one_spec.rb
@@ -6,11 +6,11 @@ describe TestReferencesOneProxy do
   include SetUpHbaseConnectionBeforeAll 
   include SetTableNamesToTestTable
 
-  let(:owner) { Person.new }
+  let(:proxy_owner) { Person.new }
   let(:target) { PersonWithTimestamp.new }
   let(:metadata) { subject.metadata }
 
-  subject { owner.send(:relation_proxy, 'boss') }
+  subject { proxy_owner.send(:relation_proxy, 'boss') }
 
   it_should_behave_like "relation proxy"
   it_should_behave_like "singular proxy"
@@ -21,13 +21,13 @@ describe TestReferencesOneProxy do
 
   describe "#find_target" do
     it "should not try to find anything if foreign key is nil" do
-      owner.boss_id = nil
+      proxy_owner.boss_id = nil
       PersonWithTimestamp.should_not_receive(:find)
       subject.load_target.should be_nil
     end
 
     it "should try to find the target if foreign key is set" do
-      owner.boss_id = "id"
+      proxy_owner.boss_id = "id"
       PersonWithTimestamp.should_receive(:find).with("id").and_return(target)
       subject.load_target.should == target
     end
@@ -38,7 +38,7 @@ describe TestReferencesOneProxy do
 
     before do
       subject.metadata.find_with = Proc.new { |target| Person.find("testing-123") }
-      owner.boss_id = nil
+      proxy_owner.boss_id = nil
     end
 
     after do
@@ -57,30 +57,30 @@ describe TestReferencesOneProxy do
   end
 
   describe "setting target object" do
-    it "should set the target's id as the foreign key in owner" do
-      owner.boss = target
-      owner.boss_id.should == target.id
+    it "should set the target's id as the foreign key in proxy_owner" do
+      proxy_owner.boss = target
+      proxy_owner.boss_id.should == target.id
     end
 
     it "should reset the targets foreign key if target is nil" do
-      owner.boss = target
-      owner.boss = nil
-      owner.boss_id.should be_nil
+      proxy_owner.boss = target
+      proxy_owner.boss = nil
+      proxy_owner.boss_id.should be_nil
     end
 
     it "should not set the target's ida as the foreign key if we are not persisting the foreign key" do
-      owner.stub(:respond_to?).and_return(false)
-      owner.boss = target
-      owner.boss_id.should be_nil
+      proxy_owner.stub(:respond_to?).and_return(false)
+      proxy_owner.boss = target
+      proxy_owner.boss_id.should be_nil
     end
 
     it "should set the target's id as the foreign key even if we are not persisting it if the record responds to setter method" do
-      owner.boss = target
-      owner.boss_id.should == target.id
+      proxy_owner.boss = target
+      proxy_owner.boss_id.should == target.id
     end
 
     it "should raise an error if there is a type mismatch" do
-      lambda { owner.boss = Person.new }.should raise_error MassiveRecord::ORM::RelationTypeMismatch
+      lambda { proxy_owner.boss = Person.new }.should raise_error MassiveRecord::ORM::RelationTypeMismatch
     end
   end